# contracts/openapi/public/identity.yaml
openapi: 3.0.3
info:
  title: COOKie Identity Public API
  version: 0.2.0
  description: >-
    Public email/password authentication and session lifecycle owned by
    Identity Service.
servers:
  - url: https://api.cookie.app/v1
    description: Production
  - url: http://127.0.0.1:8080/v1
    description: Local
tags:
  - name: auth
    description: Email/password authentication and session lifecycle.

paths:
  /auth/email/register:
    post:
      tags: [auth]
      operationId: registerWithEmail
      summary: Зарегистрироваться по email и паролю
      description: >-
        Создаёт аккаунт с неподтверждённым email и отправляет одноразовую ссылку
        подтверждения. Ответ одинаков для нового и уже зарегистрированного email,
        чтобы не раскрывать наличие аккаунта. Токены выдаются только после
        подтверждения email.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRegistrationRequest"
      responses:
        "202":
          description: Если действие допустимо, письмо подтверждения будет отправлено
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /auth/email/login:
    post:
      tags: [auth]
      operationId: loginWithEmail
      summary: Войти по email и паролю
      description: >-
        Возвращает одинаковую ошибку для неизвестного email, неверного пароля,
        неподтверждённого email и недоступного аккаунта, чтобы не позволять
        перечислять зарегистрированные адреса.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailLoginRequest"
      responses:
        "200":
          description: Токены выданы, refresh-сессия создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairWithUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Неверные учётные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /auth/email/verification/confirm:
    post:
      tags: [auth]
      operationId: confirmEmail
      summary: Подтвердить email одноразовым токеном
      description: >-
        Одноразово поглощает verification token, помечает email подтверждённым и
        создаёт первую refresh-сессию.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailVerificationConfirmRequest"
      responses:
        "200":
          description: Email подтверждён, токены выданы
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairWithUser"
        "400":
          description: Токен недействителен, уже использован или истёк
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /auth/email/verification/resend:
    post:
      tags: [auth]
      operationId: resendEmailVerification
      summary: Повторно отправить подтверждение email
      description: >-
        Всегда возвращает одинаковый ответ независимо от существования аккаунта и
        состояния email. Предыдущие неиспользованные verification tokens отзываются.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailActionRequest"
      responses:
        "202":
          description: Если действие допустимо, письмо будет отправлено
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshSession
      summary: Обновить пару токенов (refresh token rotation)
      description: >-
        Одноразово использует текущий refresh token и возвращает новую пару.
        Повторное использование уже ротированного токена считается replay:
        Identity отзывает всё семейство refresh-сессии.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Новая пара токенов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          description: Refresh token недействителен, отозван, уже использован или истёк
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Завершить одну логическую сессию устройства
      description: >-
        Любой криптографически валидный refresh token семейства аутентифицирует
        логическую сессию устройства; Identity отзывает всё семейство, включая
        токен-замену, созданный конкурентным refresh.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "204":
          description: Сессия завершена или уже была завершена
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /auth/jwks:
    get:
      tags: [auth]
      operationId: getIdentityJwks
      summary: Получить публичные ключи подписи access JWT
      description: >-
        Возвращает активный и retiring ES256 ключи для локальной проверки access JWT.
      security: []
      responses:
        "200":
          description: JSON Web Key Set
          headers:
            Cache-Control:
              schema:
                type: string
              example: public, max-age=300
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonWebKeySet"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

components:
  responses:
    BadRequest:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Слишком много запросов; повторите позднее
      headers:
        Retry-After:
          description: Число секунд до следующей разрешённой попытки
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: Сервис временно не готов обрабатывать запросы
      headers:
        Retry-After:
          description: Рекомендуемая задержка перед повтором, в секундах
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    EmailAddress:
      type: string
      format: email
      maxLength: 254
      description: >-
        Разрешены только адреса с ASCII local-part и российским доменом верхнего
        уровня .ru или .рф. Перед сравнением адрес проходит trim, Unicode NFC,
        IDNA-преобразование домена и приведение к нижнему регистру.

    Password:
      type: string
      format: password
      writeOnly: true
      minLength: 15
      maxLength: 128
      description: >-
        От 15 до 128 Unicode code points. Пробелы, Unicode whitespace/separator
        и управляющие символы запрещены; остальные требования к составу не
        навязываются. На сервере хранится только Argon2id encoded hash.

    LoginPassword:
      type: string
      format: password
      writeOnly: true
      minLength: 1
      maxLength: 128
      description: >-
        Пароль для проверки существующего credential. Текущая registration
        policy намеренно не применяется при login, чтобы её ужесточение не
        блокировало ранее созданные аккаунты.

    EmailRegistrationRequest:
      type: object
      required: [email, password]
      properties:
        email:
          $ref: "#/components/schemas/EmailAddress"
        password:
          $ref: "#/components/schemas/Password"
        locale:
          type: string
          maxLength: 35

    EmailLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          $ref: "#/components/schemas/EmailAddress"
        password:
          $ref: "#/components/schemas/LoginPassword"
        deviceId:
          type: string
          maxLength: 255

    EmailActionRequest:
      type: object
      required: [email]
      properties:
        email:
          $ref: "#/components/schemas/EmailAddress"

    EmailVerificationConfirmRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 32
          maxLength: 512
        deviceId:
          type: string
          maxLength: 255

    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 32
          maxLength: 512

    TokenPair:
      type: object
      description: >-
        Access token — короткоживущий JWT и не хранится на сервере. Refresh token —
        случайный opaque secret; клиент хранит его в Keychain/Keystore, а Identity
        сохраняет в PostgreSQL только hash и состояние rotation family.
      required: [accessToken, accessTokenExpiresIn, refreshToken, refreshTokenExpiresIn]
      properties:
        accessToken:
          type: string
        accessTokenExpiresIn:
          type: integer
          description: seconds
          example: 900
        refreshToken:
          type: string
        refreshTokenExpiresIn:
          type: integer
          description: seconds

    TokenPairWithUser:
      allOf:
        - $ref: "#/components/schemas/TokenPair"
        - type: object
          required: [user]
          properties:
            user:
              $ref: "#/components/schemas/AuthenticatedUser"

    AuthenticatedUser:
      type: object
      required: [id, newUser]
      properties:
        id:
          type: string
          format: uuid
          description: UUIDv7 account identifier.
        newUser:
          type: boolean

    JsonWebKeySet:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/JsonWebKey"

    JsonWebKey:
      type: object
      required: [kty, use, alg, kid, crv, x, y]
      properties:
        kty:
          type: string
          enum: [EC]
        use:
          type: string
          enum: [sig]
        alg:
          type: string
          enum: [ES256]
        kid:
          type: string
        crv:
          type: string
          enum: [P-256]
        x:
          type: string
        y:
          type: string

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: INVALID_REQUEST
        message:
          type: string
          example: Invalid request.
        details:
          type: object
          additionalProperties: true
        requestId:
          type: string
          format: uuid